recurso de imagem de skill adicionado

+++ src/main/java/com/ufpa/lafocabackend/api/controller/MemberController.java
@@ -179,7 +179,7 @@ public class MemberController {
 
     @CheckSecurityPermissionMethods.Level1
     @PutMapping("/{memberId}/skills/{skillId}")
-    public ResponseEntity<Void> associateSkill(@PathVariable String memberId, @PathVariable Integer skillId) {
+    public ResponseEntity<Void> associateSkill(@PathVariable String memberId, @PathVariable Long skillId) {
 
         memberService.associateSkill(memberId, skillId);
 
@@ -188,7 +188,7 @@ public class MemberController {
 
     @CheckSecurityPermissionMethods.Level1
     @DeleteMapping("/{memberId}/skills/{skillId}")
-    public ResponseEntity<Void> disassociateSkill(@PathVariable String memberId, @PathVariable Integer skillId) {
+    public ResponseEntity<Void> disassociateSkill(@PathVariable String memberId, @PathVariable Long skillId) {
 
         memberService.disassociateSkill(memberId, skillId);
